linkBasedAiAnswer/requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-qdrant
qdrant-client
openai
python-dotenv
requests

